+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- nothing yet
+### Added
+
+- Quantity field for materials & expenses line items
 
 ## [v0.1.0] - 2025-10-21